refactor: extract lastFmSeeder from queueManipulation (phase 3.4)
